Perf: Only synchronize storage on unmount.
This should save a lot of stringify / ls churn by only saving when we
absolutely must.

Additionally, I've reworked the test suite to be more modern.

This has a semantic change; state is not synced as often and we can
no longer throw errors if you have a key collision, so this will
be a major update.